refactor: #385: split withdraw and close function from position store
