+++ src/entities/position/ui/header-action-button.tsx
@@ -1,12 +1,15 @@
+import { useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import { Button } from '@penumbra-zone/ui/Button';
-import { positionsStore, DisplayPosition } from '../model/store';
+import { DisplayPosition } from '../model/store';
+import { withdrawPositions } from '../api/withdraw-positions';
+import { closePositions } from '../api/close-positions';
 
 const MAX_ACTION_COUNT = 15;
 
 export const HeaderActionButton = observer(
   ({ displayPositions }: { displayPositions: DisplayPosition[] }) => {
-    const { loading, closePositions, withdrawPositions } = positionsStore;
+    const [isLoading, setIsLoading] = useState(false);
 
     const openedPositions = displayPositions
       .filter(position => position.isOpened)
@@ -16,18 +19,6 @@ export const HeaderActionButton = observer(
         position: position.position,
       }));
 
-    if (openedPositions.length > 1) {
-      return (
-        <Button
-          actionType='destructive'
-          disabled={loading}
-          onClick={() => void closePositions(openedPositions)}
-        >
-          Close Batch ({openedPositions.length})
-        </Button>
-      );
-    }
-
     const closedPositions = displayPositions
       .filter(position => position.isClosed)
       .slice(0, MAX_ACTION_COUNT)
@@ -36,13 +27,29 @@ export const HeaderActionButton = observer(
         position: position.position,
       }));
 
+    const withdraw = async () => {
+      setIsLoading(true);
+      await withdrawPositions(closedPositions);
+      setIsLoading(false);
+    };
+
+    const close = async () => {
+      setIsLoading(true);
+      await closePositions(openedPositions);
+      setIsLoading(false);
+    };
+
+    if (openedPositions.length > 1) {
+      return (
+        <Button actionType='destructive' disabled={isLoading} onClick={() => void close()}>
+          Close Batch ({openedPositions.length})
+        </Button>
+      );
+    }
+
     if (closedPositions.length > 1) {
       return (
-        <Button
-          actionType='destructive'
-          disabled={loading}
-          onClick={() => void withdrawPositions(closedPositions)}
-        >
+        <Button actionType='destructive' disabled={isLoading} onClick={() => void withdraw()}>
           Withdraw Batch ({closedPositions.length})
         </Button>
       );
